+++ Data/SBV/Compilers/C/GMP.hs
@@ -41,7 +41,7 @@ import qualified Text.PrettyPrint.HughesPJ as P ((<>))
 
 import Data.SBV.Compilers.C.BV         (isWideBV, mappedIntegerKind)
 import Data.SBV.Compilers.C.Lowering   (CLowering, CRequirement(..), expressionLowering)
-import Data.SBV.Compilers.CodeGen      (CgConfig(..), CgSRealType(..))
+import Data.SBV.Compilers.CodeGen      (CgConfig(..))
 import Data.SBV.Core.Data
 import Data.SBV.Core.Symbolic          (NROp(..))
 
@@ -166,10 +166,6 @@ gmpExpr cfg op svs resultKind args
   = Nothing
   | IEEEFP{} <- op
   = Nothing
-  | KindCast fr to <- op
-  , Just CgLongDouble <- cgReal cfg
-  , KReal `elem` [fr, to]
-  = error "SBV->C: Exact numeric conversions involving cgSRealType CgLongDouble are not yet supported; use CgFloat, CgDouble, or exact GMP reals."
   | KindCast fr to <- op
   , Just{} <- cgReal cfg
   , KReal `elem` [fr, to]